requirements.txt
streamlit==1.29.0
opencv-python-headless==4.8.1.78
numpy>=1.24.3
pandas>=2.0.1
Pillow>=10.0.0
setuptools>=69.2.0
face_recognition==1.3.0
cmake==3.28.3
dlib==19.24.2